requirements.txt
numpy==1.17.0
matplotlib==3.3.4
torchvision==0.8.2
seaborn==0.11.1
pandas==1.1.5
ray==1.4.0
tqdm==4.61.2
torch==1.7.1
Pillow==8.3.1
scikit_learn==0.24.2
